requirements.txt
google-cloud-aiplatform>=1.70.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pillow>=10.0.0 